extract detail file components from SelectedFileViewer
