+++ src/components/File/SelectedFileviewer/InitFile.tsx
@@ -1,0 +1,53 @@
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { fileSysyemStore, PrChangedFileInfo } from "@/stores/github.store";
+
+export const InitFile = ({
+  commitFileList,
+}: {
+  commitFileList: PrChangedFileInfo[];
+}) => {
+  const setSelectedCommitFile = fileSysyemStore(
+    (state) => state.setSelectedCommitFile,
+  );
+
+  return (
+    <div className="flex w-4/5 flex-col">
+      <div className="text-2xl font-semibold">File Changed</div>
+      <Table>
+        <TableHeader>
+          <TableRow className="bg-slate-200 hover:bg-slate-200">
+            <TableHead className="w-[200px]">File Name</TableHead>
+            <TableHead>language</TableHead>
+            <TableHead>status</TableHead>
+            <TableHead className="text-right">file additions</TableHead>
+            <TableHead className="text-right">file deletions</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {commitFileList.map((file, index) => {
+            return (
+              <TableRow
+                key={index}
+                className="cursor-pointer"
+                onClick={() => setSelectedCommitFile(file)}
+              >
+                <TableCell className="font-medium">{file.filename}</TableCell>
+                <TableCell>{file.language}</TableCell>
+                <TableCell>{file.status}</TableCell>
+                <TableCell className="text-right">{file.additions}</TableCell>
+                <TableCell className="text-right">{file.deletions}</TableCell>
+              </TableRow>
+            );
+          })}
+        </TableBody>
+      </Table>
+    </div>
+  );
+};
